refactor: move step-input to own component file
